% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplot_pollen.R
\name{iplot_pollen}
\alias{iplot_pollen}
\title{Interactive Plotting Pollen Data (one season).}
\usage{
iplot_pollen(data, year)
}
\arguments{
\item{data}{A \code{data.frame} object. This \code{data.frame} should include a first column in format \code{Date} and the rest of columns in format \code{numeric} belonging to each pollen type by column.}

\item{year}{An \code{integer} value specifying the year to display. This is a mandatory argument.}
}
\value{
An interactive plot of the class \pkg{ggvis}.
}
\description{
Function to plot the pollen data during one season. The plots are fully interactive.
}
\examples{
data("munich_pollen")
iplot_pollen(data = munich_pollen, year = 2012)
}
\seealso{
\code{\link{iplot_years}}
}
