% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monthly_heatmap.R
\name{plot_monthly_heatmap}
\alias{plot_monthly_heatmap}
\title{Monthly Pollen Heatmap}
\usage{
plot_monthly_heatmap(
  data,
  pollen.type,
  location = NULL,
  location.col = NULL,
  start.month = 1,
  color.A = "yellow",
  color.B = "red",
  result = "plot"
)
}
\arguments{
\item{data}{A \code{data.frame} object. The first column must be in \code{Date}
format and pollen types must be provided as numeric columns. If the data include
several locations, one column can identify the location.}

\item{pollen.type}{A \code{character} string with the name of the pollen type to
be plotted. This name must match one numeric column in \code{data}.}

\item{location}{A \code{character} string with the location to be selected. If
\code{NULL} and a location column is detected, the first location in the data is
used. If the data do not include a location column, this argument is ignored.}

\item{location.col}{A \code{character} string with the name of the column
containing the locations. If \code{NULL}, common names such as \code{"location"},
\code{"city"} or \code{"station"} are automatically detected.}

\item{start.month}{A \code{numeric} value ranging \code{1_12} specifying the
first month shown in the x axis. The \code{start.month} argument will be
\code{1} by default.}

\item{color.A}{A \code{character} object with the specification of the color of
the lowest value for the scale. By default, \code{color.A = "yellow"}.}

\item{color.B}{A \code{character} object with the specification of the color of
the highest value for the scale. By default, \code{color.B = "red"}.}

\item{result}{A \code{character} object with the definition of the object to be
produced by the function. If \code{result == "plot"}, the function returns an
object of class \pkg{ggplot2}; if \code{result == "table"}, the function returns
a \pkg{data.frame} with the monthly sums. By default, \code{result = "plot"}.}
}
\value{
If \code{result == "plot"}, the function returns an object of class
\pkg{ggplot2}; if \code{result == "table"}, the function returns a
\pkg{data.frame} with the monthly sums.
}
\description{
Function to plot monthly sums of daily pollen concentrations as a heatmap. Months
are displayed in the x axis and years in the y axis for one selected pollen type.
}
\examples{
data("munich_pollen")
plot_monthly_heatmap(munich_pollen, pollen.type = "Betula")
plot_monthly_heatmap(munich_pollen, pollen.type = "Poaceae", start.month = 6)
}
